+++ models/campus.mjs
@@ -93,7 +93,7 @@ const driverFilter = () => ({
 CampusSchema.statics.countUsers = async function countUsers(campus, filters = {}) {
   const User = mongoose.model('User');
   const f = { ...campusFilter(campus), ...filters };
-  return User.count(f);
+  return User.countDocuments(f);
 };
 
 CampusSchema.statics.countDrivers = async function countDrivers(campus) {