+++ tests/routes/rides.test.mjs
@@ -76,7 +76,7 @@ describe('Test the rides route', () => {
         expect(response.headers['content-type']).to.contain('application/json');
       }
     } finally {
-      await Promise.all(toDropLater.map((entity) => entity.remove()));
+      await Promise.all(toDropLater.map((entity) => entity.deleteOne()));
     }
   });
 });